+++ packages/netlify-cms-core/src/lib/registry.js
@@ -1,7 +1,8 @@
 import { Map } from 'immutable';
 import produce from 'immer';
 import { oneLine } from 'common-tags';
-import EditorComponent from 'ValueObjects/EditorComponent';
+
+import EditorComponent from '../valueObjects/EditorComponent';
 
 const allowedEvents = [
   'prePublish',