+++ packages/netlify-cms-core/src/reducers/__tests__/collections.spec.js
@@ -1,4 +1,5 @@
 import { fromJS, Map } from 'immutable';
+
 import { configLoaded } from '../../actions/config';
 import collections, {
   selectAllowDeletion,
@@ -368,30 +369,15 @@ describe('collections', () => {
       });
 
       expect(selectField(collection, 'en.title')).toBe(
-        collection
-          .get('fields')
-          .get(0)
-          .get('fields')
-          .get(0),
+        collection.get('fields').get(0).get('fields').get(0),
       );
 
       expect(selectField(collection, 'it.title.subTitle')).toBe(
-        collection
-          .get('fields')
-          .get(2)
-          .get('field')
-          .get('fields')
-          .get(0),
+        collection.get('fields').get(2).get('field').get('fields').get(0),
       );
 
       expect(selectField(collection, 'fr.title.variableType')).toBe(
-        collection
-          .get('fields')
-          .get(3)
-          .get('fields')
-          .get(0)
-          .get('types')
-          .get(0),
+        collection.get('fields').get(3).get('fields').get(0).get('types').get(0),
       );
     });
   });