+++ packages/netlify-cms-core/src/reducers/index.ts
@@ -1,4 +1,5 @@
 import { List } from 'immutable';
+
 import auth from './auth';
 import config from './config';
 import integrations, * as fromIntegrations from './integrations';
@@ -13,8 +14,9 @@ import mediaLibrary from './mediaLibrary';
 import deploys, * as fromDeploys from './deploys';
 import globalUI from './globalUI';
 import status from './status';
-import { Status } from '../constants/publishModes';
-import { State, Collection } from '../types/redux';
+
+import type { Status } from '../constants/publishModes';
+import type { State, Collection } from '../types/redux';
 
 const reducers = {
   auth,